Add game layout and basic functions
Several CSS rulesets were added to the styles.css file, implementing the layout for different game elements like text, game area, controls, stats, and buttons. On the other hand, the script.js file was updated to add basic functions like going to the store, cave, fighting the dragon, buying health and weapon. This is the first step to building interactivity into the game.