Add game location objects and update function
The script has been updated to include an array of game locations, each represented as an object with properties such as name, button text, button functions, and text descriptions. An 'update' function was also created to change the game state based on the current location. The 'goStore' function has been streamlined accordingly.