Added "cave" game location and updated location transition functions
Added a new location, "cave", with corresponding interaction options to the game. Refactored the code to streamline location transition functionality: now it uses an 'update' function that adjusts the game's state based on the given location object. This change has led to optimization of the 'goStore', 'goTown', and 'goCave' functions.